moved GraphUnionProxy namespace
